Creating a new Model Instance
Added an input helper which communicates with the corresponding controller and an action which gets triggered upon clicking enter inside the input element. Creating a new record and saving it is done in the controller action that is specified along with the helper. Value option in the helper is associated with a property in the controller.