Accepting Edits
Added a new view which is registered as a helper and is instantiated when the isEditing property is set to true. The helper extends the ember textfield and takes value of that particular Todo instance's title. The acceptChanges action takes care of saving the new title provided upon clicking enter of focussing out. When the value is assigned to todo.title inside the each, the value is bound to the title name of that particular todo instance and gets set on the fly. More deep investigation is needed.