Transitioning to Show Only Incomplete Todos
The controller in the renderTemplate function generates a new todo:active controller but since the each loops over on the todo Item controller it wont affect anything. And the outlet is filled as per the template mentioned in the renderTemplate otherwise it doesnt know which to fill in it with.